More translations for user-visible strings
User-visible hungarian strings are now i18n keys.

For download statuses, instead of just the string, an object with an i18n key is sent, so that it can be translated in the renderer.
Note: download-error event left untranslated as download-error is not currently used in renderer.

Also: for temporarily invisible elements (table and status on the search page) `element.hidden = true` is used instead of `element.style.display = "none"`